tests/wpt/web-platform-tests/webvtt/parsing/cue-text-parsing/dat/tags.dat

#data
<
#errors
#document-fragment

#data
<<
#errors
#document-fragment

#data
<	
#errors
#document-fragment

#data
<

#errors
#document-fragment

#data
< 
#errors
#document-fragment

#data
<.
#errors
#document-fragment

#data
<c.
#errors
#document-fragment
| <span>

#data
</
#errors
#document-fragment

#data
<c></>x
#errors
#document-fragment
| <span>
|   "x"

#data
<c></
c>x
#errors
#document-fragment
| <span>
|   "x"
